requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-classic
chromadb
pypdf
python-dotenv
openai